Refactor orca ts_optimizier and add orca optimizer
